Rename auth type file to user and update imports
