+++ src/hooks/auth/useLoginUser.ts
@@ -2,7 +2,7 @@ import { useNavigate } from "react-router-dom";
 import { useMutation } from "@tanstack/react-query";
 import { useAuthStore } from "@/store/authStore";
 import { apiRequest } from "@/api/request";
-import { AuthResponse } from "@/types/auth";
+import { AuthResponse } from "@/types/user";
 import { LoginFormData } from "@/schemas/loginSchema";
 
 export const useLoginUser = () => {
